+++ tests/test_package.py
@@ -2,4 +2,4 @@ import phoenix_os
 
 
 def test_package_version() -> None:
-    assert phoenix_os.__version__ == "0.11.0"
+    assert phoenix_os.__version__ == "0.12.0"